pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=61",
  "wheel",
]

[project]
name = "batch-inference"
version = "1.0rc1"
description = "Batch Inference"
readme = "README.md"
authors = [{ name = "Xi Chen", email = "xichen5@microsoft.com" },
    { name = "Lu Ye", email = "luye@microsoft.com" },
    { name = "Yong Huang", email = "yohuan@microsoft.com" },
    { name = "Ze Tao", email = "zetao@microsoft.com" }]
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "filelock",
  "grpcio",
  "msgpack",
  "msgpack-numpy",
  "numpy",
]
optional-dependencies.docs = [
  "flask>=2.0.2",
  "furo>=2022.12.7",
  "sphinx>=6.1.3",
  "sphinx-autodoc-typehints!=1.23.4,>=1.22",
  "transformers>=4.27.4",
]
optional-dependencies.testing = [
  "onnxruntime",
  "pytest>=7.2.2",
  "torch",
]
[project.urls]
"Homepage" = "https://github.com/microsoft/batch-inference.git"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
